Moved the discovery code into its own thread.
This should ease some doubt over whether the chip continues to discover
over time and whether the main loop will hang indefinitely if it's not
discovering and waiting for an update of service data.

It might also be worth looking into whether non-blocking zbus connection
can accomplish the same thing.