Cleaned up the statsd output a bit.
Gathered the statsd client state into a struct, created a trait for reporting
the device state to some destination and implemented the trait for statsd.

This should make it easier for the statsd client to output to a customizable
destination and also not to have it create a new client each time it wants to
output metrics.